TPZSolutionMatrix is now an independent class
